Add files via upload
A random walker moves at discrete time 1, 2, 3, ...in a finite 2D rectangular grid either by one horizontal or a vertical step, or it does not move at all if there are no options available. 

At any discrete time $t$ and at a given point $P_t$ of the grid,  the walker randomly choose the next point $P_{t+1}$ out of a  set Neigh(P_t) . This set is made up of at most 4 points of the grid. The walker has memory, so that it remembers all the points of Neigh(P_t) which were not chosen. Those points cannot be consider as possible moves in future time